pyproject.toml
[project]
name = "article-generation"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "selenium>=4.29.0",
    "webdriver-manager>=4.0.2",
    "pyyaml>=6.0.2",
    "yandex-cloud-ml-sdk>=0.4.2",
    "numpy>=2.2.4",
    "scipy>=1.15.2",
    "psycopg2-binary>=2.9.10",
    "qdrant-client>=1.13.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/article_generation"]
